t6-common: Remove unnecessary and broad selinux permissions
The system_server denial is most likely due to TWRP creating files
under /cache/recovery, but not having the proper file_contexts to
label those files when it creates them, so they're "unlabelled". You
should be able to remedy this denial by running restorecon on
/cache/recovery.

The priv_app denial is too broad. I haven't figured out how to resolve
this Binder denial, but also haven't seen any issues when it is
denied. Leave it alone unless a more specific rule can be made.

Change-Id: Ibf3458022955c3b7ef23206cbeb01137d042728b